+++ pf1.js
@@ -152,7 +152,7 @@ Hooks.once("setup", function() {
  */
 Hooks.once("ready", async function() {
   // Migrate data
-  const NEEDS_MIGRATION_VERSION = "0.75.6";
+  const NEEDS_MIGRATION_VERSION = "0.75.9";
   let PREVIOUS_MIGRATION_VERSION = game.settings.get("pf1", "systemMigrationVersion");
   if (typeof PREVIOUS_MIGRATION_VERSION === "number") {
     PREVIOUS_MIGRATION_VERSION = PREVIOUS_MIGRATION_VERSION.toString() + ".0";